+++ pkg/analyzer/filter_fields.go
@@ -587,8 +587,8 @@ var (
 
 	// FieldWeekday is a query result column.
 	FieldWeekday = Field{
-		querySessions:  "toDayOfWeek(time, 1, '%s')",
-		queryPageViews: "toDayOfWeek(time, 1, '%s')",
+		querySessions:  "toDayOfWeek(time, %d, '%s')",
+		queryPageViews: "toDayOfWeek(time, %d, '%s')",
 		queryDirection: "ASC",
 		queryWithFill:  "WITH FILL FROM 0 TO 7",
 		timezone:       true,