+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * added non-interactive events
 * added event meta keys to filter options
 * added browser and OS to filter options
+* added weekday mode configuration to filter
 * removed the requirement for event to be set to filter for key/value options
 * updated dependencies